transaction/history: initial history ingestion system
Introduces a history ingestion system that supports
multiple blockchain sources. The system allows for
asynchronous processing of history blocks in order
drastically increase the replay speed of history.

Ingestion has been designed in such that it is safe
for multiple ingestors to fight over ingestion, but
aims to run a single instance of ingestion per
ingestor in order to avoid duplicate work.

GitOrigin-RevId: 00e39b867f95928d1b6ab3eefb711bb1e23938c0