transaction: agora history read support.
This changes shifts the reading of transactions
from Horizon directly to the Agora history store.

For explicit transaction lookups, the server has
the ability to fallback to Horizon to increase
lookup responsiveness and availability.

For GetHistory, requests cannot fallback to horizon,
as the underlying horizon instance may not have enough
history to provide a complete view from when the client
last synced. This would result in the client updating
their cursor, and never knowing to retry looking back
in time.

GitOrigin-RevId: 29fa316234210e4c4d0401fea9086d37c37dad1d