transaction/server: do not call SignTransaction() webhook on earns.
As earns are considered 'app' -> 'user', and the 'app'
controls the webhook, it makes little sense to call the
the SignTransaction webhook on Earn transactions.

Furthermore, it's quite likely that the Earn sender
will be the same as the whitelist key (as developers
don't always have a dedicated whitelist account). In
this case, the following safety check will cause the
earn to fail:

if payment.Sender() == whitelistKey.Public() {
    resp.Reject()
}

If the whitelist key contains Kin (which is the case
if the hot wallet is also the whitelist wallet), then
an attack could craft a transaction that can steal Kin
the whitelist account. Not forwarding earns alleviates
this case.

GitOrigin-RevId: e5d2879a47e6e38d6037d8604bfbff59df0d0718