client: use hacky string conversion instead of big.Float
Using big.Float and friends results in some values of Kin
not being able to be accurately represented. Since the Kin
to quark conversion is fixed, and we don't have to do any
arithmetic operations on Kin values, we can exploit the
string representation to convert between Kin and Quarks.

GitOrigin-RevId: 50e56b1b54c8384bc19c8e9ecafea7525a847193