service/agora: partition solana clients.
The internal RPC nodes seem to be suffering heavily,
so we partition the use of endpoints based on urgency.

The public endpoints appear to be more stable, but have
rate limits. As a large portion of our requests is for the
ingestors, we keep the ingestion on our local cluster.

GitOrigin-RevId: bf6219b8913456d0c6e9208aab5232dc2a8c4928