account: add reverse mapper.
This should help us identify later which accounts
requested are token accounts, and what are completely
unknown.

This differentiation can help us make a decision for whether
or not we expect the account to exist or not. This will
allow us to then heavily cache / prevent requests to the
solana RPC node.

GitOrigin-RevId: 0f9b4d7c903bda2df4e19defeb0a24bb54ba1191