transaction: speculative balance updates.
Currently, there is significant delay from the
time a transaction is submitted to the time it
is observed. This could be related to our nodes
being separate, or slightly slower, but it's
unclear.

However, for the end users that actively observe
and care about their balances, we can sepculatively
cache what we predict the balance to be, if they're
using agora.

We do this by loading the balance before a transaction,
and then updating it (in cache, with a somewhat short
ttl) to what we predict it would be, based on a
successful transaction. This should help GetAccount()
look more live for users who are not sending a lot of
concurrent transactions. That is, non-earn wallets
should benefit from this greatly.

GitOrigin-RevId: 1bc1584ca86bdcbdae5e3b78453692a9ef7e8ef7