events: add a redis based event bus for transaction events.
Previously, each node would watch the blockchain
directly for transactions, triggering events for
any active stream. However, with the introduction
of speculative execution, and the fact that the
ConfirmedBlock API lags significantly (due to
it requiring rooted blocks), the event stream system
became wildly out of sync with what other RPCs would
yield.

For example, a SubmitTransaction() call would be
reflected almost immediately with a GetAccountInfo(),
but no event (transaction or balance change) would be
sent to listeners for almost 30 seconds (in the happy
path).

Since the SubmitTransaction() call can occur on a
different server than the account listener, we must
introduce a shared event bus for notifications to
work properly.

Redis works well at the current scale, as the number
of Agora triggered events is fairly low, as well as
the number of Agora servers.

Further work will  be done to have the history
collector listen to ConfirmedBlocks, and trigger
events on the EventBus, either introducing a
'confirmed' flag, or only triggering events that are
not sent by Agora.

GitOrigin-RevId: adda292a098422a804dab47c0fc50cc79d9f1574