transaction/history: switch GetTransactions() query params.
Previously, GetTransactions() used blocks as query
parameters. This was based on the assumption that
entire blocks were committed at a time. However,
this level of granularity is insufficient for
proper paging logic.

For example, suppose that two blocks A and B
had a size of 200 and 2000 respectively. A query
starting at A with a limit of 1000 would yield
all the entries from A, and 800 entries of B.
The next interation in the paging would either have
to reprocess B again, reprocessing the first 800.

Furthermore, in the above example, the logic would
never be able to fully process B without increasing
its limit.

This change allows querying at the entry ordering
key level, allowing for precise querying and paging
of transaction history.

GitOrigin-RevId: 7d12c9f3f80a314598cbcfc62fafb45848a6b8f9