migration/processor: gracefully handle burned accounts.
In the vast majority of cases, accounts that were
burned on kin2 were migrated to kin3, and so the
effective result from the dual migrator would result
in an actual account.

However, there are cases where an account was burned
on kin2, and not migrated. In this case, the dual
migrator would yield ErrNotFound on the new chain. This
shouldn't be considered an error.

GitOrigin-RevId: 707685329a520e577a5fa6d30b5cc71b01f22ef7