account/info: remove account store.
Currently we only use the account cache.

The store was originally intended to replace having
to query the underlying blockchain (directly), allowing
for better MVVC semantics for speculative states.

However, the underlying blockchain has since become a lot more
stable, and the need for a stronger store than a cache is no
longer required.

GitOrigin-RevId: 3fec2da95a646905a936deb1a4c4202061569214